Mengganti kode model lama
Mengganti kode yang digunakan untuk melatih model lama dengan yang baru
